新規登録処理とログイン処理を分割
バリデーションの修正
Formフィールド変更
