+++ src/main/java/com/example/bugbug/controller/IndexController.java
@@ -3,20 +3,22 @@ package com.example.bugbug.controller;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import javax.servlet.http.HttpSession;
 
-// マイアカウント関係のコントローラー
 @RequiredArgsConstructor
 @Controller
-public class MyAccountController {
+public class IndexController {
 
     @Autowired
     private HttpSession session;
 
-    @RequestMapping("logout")
-    public String logout(){
+    @RequestMapping(value = {"/", "/index", "/index.html"})
+    public String viewIndex(Model model){
+        model.addAttribute("session_name",session.getAttribute("user_name"));
         return "index";
     }
 }
