+++ src/main/java/com/example/bugbug/validator/MailValidator.java
@@ -2,13 +2,13 @@ package com.example.bugbug.validator;
 
 import java.util.List;
 
+import com.example.bugbug.service.AccountService;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import org.springframework.validation.Validator;
 
 import com.example.bugbug.entity.Users;
 import com.example.bugbug.form.InputForm;
-import com.example.bugbug.service.UserService;
 
 import lombok.RequiredArgsConstructor;
 //メールのバリデーション
@@ -16,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 @Component
 public class MailValidator implements Validator {
 	
-	private final UserService service;
+	private final AccountService service;
 
 	@Override
 	public boolean supports(Class<?> clazz) {
@@ -29,6 +29,8 @@ public class MailValidator implements Validator {
 		InputForm form = (InputForm) target;
 		//メールの検索
 		List<Users> list = service.findMail(form.getMail());
+		System.out.println(list);
+		System.out.println("validate call");
 		//メールがすでに登録されている場合
 		if(!list.isEmpty()) {
 			errors.reject("com.example.demo.validator.MailValidator.message");
