+++ src/main/resources/templates/index.html
@@ -7,9 +7,10 @@
 <body>
 
 <h2>トップページ</h2>
-<form th:action="@{/login/form}" method="post">
-	<input type="submit" value="ログイン" > 
-</form>
+<a href="/login/form">ログイン</a>
+<a href="/sinup/form">新規登録</a>
+<a href="/logout">ログアウト</a>
+<p th:text="${session_name}+さんようこそ"></p>
 <th:block th:if="${msg} != null">
 	[[${msg}]]
 </th:block>
