+++ src/main/resources/templates/sinup.html
@@ -16,11 +16,11 @@
 	</div>
 	<div>
 		パスワード<span>必須</span><br>
-		<input type="text" name="pass" th:field="*{pass}"> 
+		<input type="text" name="pass" th:field="*{password}">
 	</div>
 	<div>
 		パスワード（確認用）<span>必須</span><br>
-		<input type="text" name="comPass" th:field="*{Pass}">
+		<input type="text" name="comPass" th:field="*{confirmPassword}">
 	</div>
 	<div>
 		ユーザーネーム<span>必須</span><br>
